CORE-15703 Moved certificate initialization to start of @BeforeAll
